backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-environ>=0.11.0
psycopg2-binary>=2.9.9
celery>=5.3.4
redis>=5.0.1
channels>=4.0.0
channels-redis>=4.1.0
supabase>=2.0.0
boto3>=1.29.0
google-generativeai>=0.3.0
assemblyai>=0.17.0
Pillow>=10.1.0
python-multipart>=0.0.6
gunicorn>=21.2.0
uvicorn[standard]>=0.25.0
websockets>=12.0
